Refactor:  ReportDetail state, sideEffect를 contract 패키지로 이동
