+++ presentation/src/main/java/com/threegap/bitnagil/presentation/reportdetail/contract/ReportDetailSideEffect.kt
@@ -1,0 +1,3 @@
+package com.threegap.bitnagil.presentation.reportdetail.contract
+
+interface ReportDetailSideEffect
