--- presentation/src/main/java/com/threegap/bitnagil/presentation/reportdetail/model/mvi/ReportDetailSideEffect.kt
@@ -1,3 +1,0 @@
-package com.threegap.bitnagil.presentation.reportdetail.model.mvi
-
-interface ReportDetailSideEffect
